requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile ./requirements.in
#
attrs==22.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-aws-glue-alpha==2.38.0a0
    # via -r ./requirements.in
aws-cdk-aws-lambda-python-alpha==2.38.0a0
    # via -r ./requirements.in
aws-cdk-lib==2.38.0
    # via
    #   -r ./requirements.in
    #   aws-cdk-aws-glue-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   cdk-nag
cattrs==22.1.0
    # via jsii
cdk-nag==2.15.46
    # via -r ./requirements.in
constructs==10.1.78
    # via
    #   -r ./requirements.in
    #   aws-cdk-aws-glue-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
exceptiongroup==1.0.0rc8
    # via cattrs
jsii==1.64.0
    # via
    #   aws-cdk-aws-glue-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-aws-glue-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via
    #   -c ./pod_metadata_extractor/runtime/requirements.txt
    #   jsii
six==1.16.0
    # via
    #   -c ./pod_metadata_extractor/runtime/requirements.txt
    #   python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-aws-glue-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
typing-extensions==4.3.0
    # via jsii
